refactor: move decode function and schema init to adapters + inject on kubegraph
